Optimize readInt for values < 16 bits by processing an entire 32-bit int in a single chunk, remove Ints.checkedCasts

+++ pinot-core/src/main/java/com/linkedin/pinot/core/index/reader/impl/FixedBitSkipListSCMVReader.java
@@ -167,7 +167,7 @@ public class FixedBitSkipListSCMVReader implements SingleColumnMultiValueReader 
     return docsPerChunk;
   }
 
-  private int computeLength(long rowOffSetStart) {
+  private int computeLength(int rowOffSetStart) {
     long rowOffSetEnd = customBitSet.nextSetBitAfter(rowOffSetStart);
     if (rowOffSetEnd < 0) {
       return (int) (totalNumValues - rowOffSetStart);
